refactored getting blog posts into cached function
also tried implemented a dynamic sitemap file
