+++ app/posts/page.tsx
@@ -1,29 +1,19 @@
-import { BlogFactoryNextJS } from "@blogfactory/nextjs";
 import Link from "next/link";
 import BlogPostCardPreview from "@/components/blog-post-card-preview";
 import GoBackButton from "@/components/go-back-button";
 import { buttonVariants } from "@/components/ui/button";
+import { getBlogFactoryPosts } from "@/lib/blog-factory";
 
 export default async function PostsPage() {
-  const apiKey = String(process.env.BLOG_FACTORY_API_KEY);
-  const blogFactory = new BlogFactoryNextJS(apiKey).app;
+  const posts = await getBlogFactoryPosts();
 
-  const { data } = await blogFactory.listPosts();
-
-  if (!data) {
+  if (!posts.length) {
     return (
       <div>
         <h1>No Posts Found</h1>
       </div>
     );
   }
-
-  const posts = data.data.slice(0, 3).map((post) => ({
-    title: post.title,
-    urlSlug: post.urlSlug,
-    content: post.content,
-    created: post.publishedAt,
-  }));
 
   return (
     <div className="p-8">
@@ -41,11 +31,11 @@ export default async function PostsPage() {
       </div>
       <ul className="grid grid-cols-1 md:grid-cols-2 xl:grid-cols-3 gap-4">
         {posts
+          .slice(0, 3)
           .sort(
             (a, b) =>
               new Date(b.created).getTime() - new Date(a.created).getTime(),
           )
-          .slice()
           .map((post) => (
             <BlogPostCardPreview
               key={post.title}
