+++ app/sitemap.ts
@@ -1,18 +1,23 @@
 import { MetadataRoute } from "next";
+import { getBlogFactoryPosts } from "@/lib/blog-factory";
+
+const URL = "https://feedback-loop.io";
 
 export default function sitemap(): MetadataRoute.Sitemap {
-  return [
-    {
-      url: "https://feedback-loop.io",
-      lastModified: new Date(),
-      changeFrequency: "weekly",
-      priority: 1,
-    },
-    {
-      url: "https://feedback-loop.io/posts",
-      lastModified: new Date(),
-      changeFrequency: "weekly",
-      priority: 0.8,
-    },
-  ];
+  let posts: MetadataRoute.Sitemap = [];
+
+  getBlogFactoryPosts().then((response) =>
+    response.map((post) => ({
+      url: `${URL}/posts/${post.urlSlug}`,
+      lastModified: new Date().toISOString(),
+    })),
+  );
+
+  const routes: MetadataRoute.Sitemap = ["/", "/posts"].map((route) => ({
+    url: `${URL}${route}`,
+    lastModified: new Date().toISOString(),
+    changeFrequency: "weekly" as const,
+  }));
+
+  return [...posts, ...routes];
 }
